backend/requirements.txt
fastapi
uvicorn
plaid-python
python-dotenv
httpx
pycryptodome
